Put back DateTimeUtils and enrich SparkDateTimeUtils.
