+++ connector/connect/common/src/main/scala/org/apache/spark/sql/connect/common/LiteralValueProtoConverter.scala
@@ -32,7 +32,7 @@ import com.google.protobuf.ByteString
 
 import org.apache.spark.connect.proto
 import org.apache.spark.sql.catalyst.ScalaReflection
-import org.apache.spark.sql.catalyst.util.{DateTimeUtils, IntervalUtils}
+import org.apache.spark.sql.catalyst.util.{IntervalUtils, SparkDateTimeUtils}
 import org.apache.spark.sql.connect.common.DataTypeProtoConverter._
 import org.apache.spark.sql.types._
 import org.apache.spark.unsafe.types.CalendarInterval
@@ -88,10 +88,10 @@ object LiteralValueProtoConverter {
       case v: LocalDate => builder.setDate(v.toEpochDay.toInt)
       case v: Decimal =>
         builder.setDecimal(decimalBuilder(Math.max(v.precision, v.scale), v.scale, v.toString))
-      case v: Instant => builder.setTimestamp(DateTimeUtils.instantToMicros(v))
-      case v: Timestamp => builder.setTimestamp(DateTimeUtils.fromJavaTimestamp(v))
-      case v: LocalDateTime => builder.setTimestampNtz(DateTimeUtils.localDateTimeToMicros(v))
-      case v: Date => builder.setDate(DateTimeUtils.fromJavaDate(v))
+      case v: Instant => builder.setTimestamp(SparkDateTimeUtils.instantToMicros(v))
+      case v: Timestamp => builder.setTimestamp(SparkDateTimeUtils.fromJavaTimestamp(v))
+      case v: LocalDateTime => builder.setTimestampNtz(SparkDateTimeUtils.localDateTimeToMicros(v))
+      case v: Date => builder.setDate(SparkDateTimeUtils.fromJavaDate(v))
       case v: Duration => builder.setDayTimeInterval(IntervalUtils.durationToMicros(v))
       case v: Period => builder.setYearMonthInterval(IntervalUtils.periodToMonths(v))
       case v: Array[_] => builder.setArray(arrayBuilder(v))
@@ -263,13 +263,13 @@ object LiteralValueProtoConverter {
       case proto.Expression.Literal.LiteralTypeCase.STRING => literal.getString
 
       case proto.Expression.Literal.LiteralTypeCase.DATE =>
-        DateTimeUtils.toJavaDate(literal.getDate)
+        SparkDateTimeUtils.toJavaDate(literal.getDate)
 
       case proto.Expression.Literal.LiteralTypeCase.TIMESTAMP =>
-        DateTimeUtils.toJavaTimestamp(literal.getTimestamp)
+        SparkDateTimeUtils.toJavaTimestamp(literal.getTimestamp)
 
       case proto.Expression.Literal.LiteralTypeCase.TIMESTAMP_NTZ =>
-        DateTimeUtils.microsToLocalDateTime(literal.getTimestampNtz)
+        SparkDateTimeUtils.microsToLocalDateTime(literal.getTimestampNtz)
 
       case proto.Expression.Literal.LiteralTypeCase.CALENDAR_INTERVAL =>
         new CalendarInterval(
